Revert "Bug 1181572 - Stop retrieving rows of type 'pending' from job_eta table"
Since for whatever reason this increased the time for the query rather
than reducing it, even though the EXPLAIN says half the rows are now
being scanned. Let's just revert this part and wait for the job_eta
rewrite in bug 1196764 rather than spend any more time on query
optimisation.

This reverts commit 521883a9df59b46c7ba5b5fac009542948a3a637.